requirements.txt
Django==3.2.6
django-apscheduler==0.6.0
feedparser==6.0.8
python-dateutil==2.8.2
psycopg2==2.9.6
beautifulsoup4==4.12.2